requirements.txt
# 仅允许在远程环境执行安装，例如 `pip install -r requirements.txt`
accelerate>=0.30,<0.35
bitsandbytes>=0.43,<0.45
datasets>=2.18,<3.0
evaluate>=0.4,<0.6
huggingface_hub>=0.23,<0.25
jinja2>=3.1,<3.2
lm-eval>=0.4,<0.5
numpy>=1.24,<2.0
peft>=0.11,<0.12
pyyaml>=6.0,<7.0
rich>=13.7,<14.0
tokenizers>=0.15,<0.20
torch>=2.2,<2.4
transformers>=4.43,<5.0
trl>=0.9,<0.11
